requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
distro==1.9.0
dnspython==2.8.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==2.2.6
openai==2.8.1
pandas==2.3.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pymongo==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
